Fix stubbing failure
