+++ pom.xml
@@ -19,25 +19,15 @@
     </testResources>
     <plugins>
       <plugin>
-        <groupId>org.codehaus.gmaven</groupId>
-        <artifactId>gmaven-plugin</artifactId>
+        <groupId>org.codehaus.gmavenplus</groupId>
+        <artifactId>gmavenplus-plugin</artifactId>
         <version>1.5</version>
-        <extensions>true</extensions>
         <executions>
           <execution>
             <goals>
+              <goal>compile</goal>
               <goal>testCompile</goal>
             </goals>
-            <configuration>
-              <sources>
-                <fileset>
-                  <directory>${pom.basedir}/src/test/groovy</directory>
-                  <includes>
-                    <include>**/*.groovy</include>
-                  </includes>
-                </fileset>
-              </sources>
-            </configuration>
           </execution>
         </executions>
       </plugin>
@@ -69,13 +59,19 @@
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
-      <version>4.12</version>
+      <version>4.11</version>
       <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>org.spockframework</groupId>
       <artifactId>spock-core</artifactId>
       <version>1.0-groovy-2.4</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>cglib</groupId>
+      <artifactId>cglib-nodep</artifactId>
+      <version>2.2</version>
       <scope>test</scope>
     </dependency>
